requirements.txt
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
pyyaml==6.0.1
